ADDED RESIZABLE CSS LOGO
yo guys the css logo Chase exported it seems he exported its container. As a result i cannot resize it as the screen enlarges hence i have deleted it and added the actual logo which is resizable. So you guys should delete your css logo file  in the Images folder as it may conflict with the one here already when we pull request.